Proposal for patching SonarQube rule 2184

+++ pulsar-client/src/main/java/org/apache/pulsar/client/impl/conf/ConsumerConfigurationData.java
@@ -90,7 +90,7 @@ public class ConsumerConfigurationData<T> implements Serializable, Cloneable {
     
     private boolean autoAckOldestChunkedMessageOnQueueFull = false;
 
-    private long expireTimeOfIncompleteChunkedMessageMillis = 60 * 1000;
+    private long expireTimeOfIncompleteChunkedMessageMillis = (long) 60 * 1000;
 
     @JsonIgnore
     private CryptoKeyReader cryptoKeyReader = null;
